factors out test_asr
